Better wording in the prompt message
VK API requires to enter several digits from a phone number of a user
when it wants to make an addtional check for some security reason.
The user is required to enter only masked digets from his phone number,
but this is not clear from the promt message.

Adjust the text of the message to clearly show what digits
should be entered